# src/main/java/be/ipam/cryptowallet/api/WalletApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package be.ipam.cryptowallet.api;

import be.ipam.cryptowallet.api.model.Wallet;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@Validated
@Api(value = "Wallet", description = "the Wallet API")
public interface WalletApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/v1/wallets/ : Ajouter un wallet
     * Ajout un wallet
     *
     * @param wallet Wallet object (optional)
     * @return created (status code 201)
     */
    @ApiOperation(value = "Ajouter un wallet", nickname = "addWallet", notes = "Ajout un wallet", authorizations = {
        @Authorization(value = "bearerAuth")
    }, tags={ "Wallet", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "created") })
    @RequestMapping(value = "/api/v1/wallets/",
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Wallet> addWallet(@ApiParam(value = "Wallet object"  )  @Valid @RequestBody(required = false) Wallet wallet) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/v1/wallets/{walletId} : Delete wallet
     * Delete user
     *
     * @param walletId Wallet Identifier (required)
     * @return No content (status code 204)
     */
    @ApiOperation(value = "Delete wallet", nickname = "deleteWalletById", notes = "Delete user", authorizations = {
        @Authorization(value = "bearerAuth")
    }, tags={ "Wallet", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "No content") })
    @RequestMapping(value = "/api/v1/wallets/{walletId}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deleteWalletById(@ApiParam(value = "Wallet Identifier",required=true) @PathVariable("walletId") long walletId) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/wallets/{walletId} : Retourne un wallet
     * Retourne un wallet
     *
     * @param walletId Wallet Identifier (required)
     * @return successful operation (status code 200)
     *         or Given user ID doesn&#39;t exist (status code 404)
     */
    @ApiOperation(value = "Retourne un wallet", nickname = "getWalletById", notes = "Retourne un wallet", response = Wallet.class, responseContainer = "List", authorizations = {
        @Authorization(value = "bearerAuth")
    }, tags={ "Wallet", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Wallet.class, responseContainer = "List"),
        @ApiResponse(code = 404, message = "Given user ID doesn't exist") })
    @RequestMapping(value = "/api/v1/wallets/{walletId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Wallet> getWalletById(@ApiParam(value = "Wallet Identifier",required=true) @PathVariable("walletId") long walletId) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"walletId\" : 0.8008281904610115, \"addressBtc\" : \"addressBtc\", \"name\" : \"name\", \"addressEth\" : \"addressEth\", \"transactions\" : [ { \"date\" : \"date\", \"number\" : 1.4658129805029452, \"price\" : 5.962133916683182, \"operation\" : 5.637376656633329, \"transactionId\" : 6.027456183070403, \"crypto\" : { \"symbol\" : \"symbol\", \"name\" : \"name\", \"cryptoId\" : 2.3021358869347655, \"lastPrice\" : 7.061401241503109 } }, { \"date\" : \"date\", \"number\" : 1.4658129805029452, \"price\" : 5.962133916683182, \"operation\" : 5.637376656633329, \"transactionId\" : 6.027456183070403, \"crypto\" : { \"symbol\" : \"symbol\", \"name\" : \"name\", \"cryptoId\" : 2.3021358869347655, \"lastPrice\" : 7.061401241503109 } } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /api/v1/wallets/{walletId} : update the wallet
     * update the wallet
     *
     * @param walletId Wallet Identifier (required)
     * @param wallet Wallet object (optional)
     * @return Updated (status code 200)
     */
    @ApiOperation(value = "update the wallet", nickname = "updateWallet", notes = "update the wallet", authorizations = {
        @Authorization(value = "bearerAuth")
    }, tags={ "Wallet", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Updated") })
    @RequestMapping(value = "/api/v1/wallets/{walletId}",
        consumes = { "application/json" },
        method = RequestMethod.PATCH)
    default ResponseEntity<Void> updateWallet(@ApiParam(value = "Wallet Identifier",required=true) @PathVariable("walletId") long walletId,@ApiParam(value = "Wallet object"  )  @Valid @RequestBody(required = false) Wallet wallet) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}